[PHOENIX-7987]-Code refactoring
