+++ egov/egov-ptisweb/src/main/webapp/WEB-INF/layout/tiles.xml
@@ -580,7 +580,7 @@
     </definition>
     <definition name="searchSurveyApplication-form" extends="base_layout">
         <put-attribute name="page-title" value="lbl.search.survey.application" cascade="true"/>
-        <put-attribute name="body" value="/WEB-INF/views/es/searchSurveyApplication-form.jsp"/>
+        <put-attribute name="body" value="/WEB-INF/views/transactions/survey/searchSurveyApplication-form.jsp"/>
     </definition>
     <definition name="surveyApplication-form" extends="base_layout">
         <put-attribute name="page-title" value="lbl.search.survey.application" cascade="true"/>
